Add hooks to dbt seeds
